requirements.txt
Django==5.2.7
djangorestframework==3.16.1
Pillow==12.0.0
python-decouple==3.8
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
